visit shop bug fixed

+++ src/main/java/com/ecom/dohsaround/CustomerController/CartController.java
@@ -1,6 +1,5 @@
 package com.ecom.dohsaround.CustomerController;
 
-
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
@@ -35,22 +34,22 @@ public class CartController {
         }
         String username = principal.getName();
         Customer customer = customerService.findByUsername(username);
-        ShoppingCart shoppingCart = customer.getShoppingCart();
-        if (shoppingCart == null) {
-            model.addAttribute("check", "No item in your cart");
+        if (customer != null) {
+            ShoppingCart shoppingCart = customer.getShoppingCart();
+            if (shoppingCart == null) {
+                model.addAttribute("check", "No item in your cart");
+            }
+
+            model.addAttribute("principal", principal);
+            model.addAttribute("customer", customer);
+
+            // session.setAttribute("totalItems", shoppingCart.getTotalItems());
+            // model.addAttribute("subTotal", shoppingCart.getTotalPrices());
+            model.addAttribute("shoppingCart", shoppingCart);
         }
-
-        model.addAttribute("principal", principal);
-        model.addAttribute("customer", customer);
-
-       // session.setAttribute("totalItems", shoppingCart.getTotalItems());
-      //  model.addAttribute("subTotal", shoppingCart.getTotalPrices());
-        model.addAttribute("shoppingCart", shoppingCart);
-
 
         return "cart_main";
     }
-
 
     @PostMapping("/add-to-cart")
     public String addItemToCart(
@@ -65,51 +64,58 @@ public class CartController {
         Product product = productService.getProductById(productId);
         String username = principal.getName();
         Customer customer = customerService.findByUsername(username);
+        if (customer != null) {
 
         ShoppingCart cart = cartService.addItemToCart(product, quantity, customer);
+        }else{
+            return "redirect:/login";
+        }
         return "redirect:" + request.getHeader("Referer");
 
     }
 
-
     @RequestMapping(value = "/update-cart", method = RequestMethod.POST, params = "action=update")
     public String updateCart(@RequestParam("quantity") int quantity,
-                             @RequestParam("id") Long productId,
-                             Model model,
-                             Principal principal
-    ) {
+            @RequestParam("id") Long productId,
+            Model model,
+            Principal principal) {
 
         if (principal == null) {
             return "redirect:/login";
         } else {
             String username = principal.getName();
             Customer customer = customerService.findByUsername(username);
+            if (customer != null) {
             Product product = productService.getProductById(productId);
             ShoppingCart cart = cartService.updateItemInCart(product, quantity, customer);
 
             model.addAttribute("shoppingCart", cart);
+            }
             return "redirect:/cart";
         }
 
     }
 
-
     @RequestMapping(value = "/update-cart", method = RequestMethod.POST, params = "action=delete")
     public String deleteItemFromCart(@RequestParam("id") Long productId,
-                                     Model model,
-                                     Principal principal, HttpServletRequest request) {
+            Model model,
+            Principal principal, HttpServletRequest request) {
         if (principal == null) {
             return "redirect:/login";
         } else {
             String username = principal.getName();
             Customer customer = customerService.findByUsername(username);
+            if (customer != null) {
             Product product = productService.getProductById(productId);
             ShoppingCart cart = cartService.deleteItemFromCart(product, customer);
             model.addAttribute("shoppingCart", cart);
+            
             return "redirect:" + request.getHeader("Referer");
+            }else{
+                return "redirect:/login";
+            }
         }
 
     }
-
 
 }
